chore: restore MessageItem import indentation
